pyproject.toml
[tool.poetry]
name = "finance-helper"
version = "0.1.0"
description = ""
authors = ["Junnoh Lee <pluruel@gmail.com>"]
readme = "README.md"
packages = [{include = "finance_helper"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.94.1"
uvicorn = "^0.21.0"
alembic = "^1.10.2"
pydantic = {extras = ["email"], version = "^1.10.6"}
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.5"
strawberry-graphql = {extras = ["fastapi"], version = "^0.165.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
